+++ src/MBC3.h
@@ -3,13 +3,13 @@
 class MBC3 : public Memory
 {
 public:
-	MBC3(uint8_t* rom, uint8_t romBanks, uint8_t ramBanks, uint8_t* exram);
+	MBC3(uint8_t* rom, uint16_t romBanks, uint8_t ramBanks, uint8_t* exram);
 	~MBC3();
 	uint8_t read8(uint16_t add);
 	void write8(uint16_t add, uint8_t val);
 
 	uint8_t* exram = nullptr;
-	uint8_t romBanks = 0;
+	uint16_t romBanks = 0;
 	uint8_t ramBanks = 0;
 
 	uint8_t RAMG = 0;
